refactor: split ManualInputForm
