+++ src/features/checks/manual-input-form/styles.ts
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+import { Div, Text } from '@vkontakte/vkui';
+
+export const StyledForm = styled.form`
+  text-align: left;
+`;
+
+export const CenteredDiv = styled(Div)`
+  text-align: center;
+`;
+
+export const CenteredTitle = styled(Text)`
+  text-align: center;
+  margin-bottom: 12px;
+`;
